Migrate radix sort
- Move implementation to header file
- Add missing curly braces
- Update README
- Delete old quick_sort.cpp

Note that since the implementation of this algorithm can only sort on
positive values, I'm leaving off integrating unit tests for now. Once we
 have an implementation that can sort on negative values it'll be very
 easy to add it to the unit test file (sorting.cpp).
