+++ C++/README.md
@@ -36,7 +36,7 @@ Implementation of well-known (and some rare) algorithms, in C++.
   * [Insertion sort](include/algorithm/sorting/insertion_sort.hpp) :white_check_mark:
   * [Merge sort](include/algorithm/sorting/merge_sort.hpp) :white_check_mark:
   * [Quick sort](include/algorithm/sorting/quick_sort.hpp) :white_check_mark:
-  * [Radix sort](source/sorting/radix_sort.cpp)
+  * [Radix sort](include/algorithm/sorting/radix_sort.hpp) 
   * [Selection sort](source/sorting/selection_sort.cpp)
   * [Shell sort](source/sorting/shell_sort.cpp)
 * String
